Resolved issue #471 - Locked chests can't be opened:
Locked chests were unable to be looted as after an attempt to open them, they
would quickly close and the LootView would never appear.

This however was not the case with locked chests that required lockpicking.
Postboxes in Stratholme still remain broken.

Thanks to LordUsagi for helping me out with this issue.