Fixed creature AI scripts:
Scripts that casted spells on players that were in range on creatures death were not working.
After hours of searching and testing, and a lot of help from LordUsagi this was resolved in two steps.

1) Script entries in "creature_ai_scripts" that fall in the category explained above
   had target parameter (action1_param2 field in DB) I believe by mistake set to
   TARGET_T_HOSTILE instead of TARGET_T_SELF. I think this needs to be set like this
   because if a dead caster tries to cast a spell on a hostile target, the server will
   find none because the caster is not in combat and cannot be in combat when he's dead.

2) The scripts also have castFlag set to TARGET_UNIT_AREA_ENTRY_SRC, which means the spell
   is some kind of an AOE spell. The grid would then be searched for all hostile targets,
   again only to find none because the 'isFriendlyTo' and 'isHostileTo' both return their respected
   equivalent of not hostile. There were only two ways to find out if a target was hostile towards
   a dead caster. One of them is to check if the caster is in combat with target, and the other
   is a faction check which leaves all neutral creatures unable to cast AOE spells on death.

I solved this by implementing a feature that resembles unit's memory. Every non-player unit will
store their attackers in a special list, and can remember these entries even after death.

The list is cleared a tick or two before the corpse is removed from the map.