Resolved issue #261:
It was reported that Shadowmeld can be cast right after casting an auto-shoot spell,
and this would result in the caster being successfully under the effect of stealth while
continually attacking his victim.

The culprit behind this issue is a lack of middle phase between the spell being cast
and hitting it's target, and the player would only be flagged in combat once his victim
received initial damage. This created a window of opportunity for a crafty player to cast
a spell that normally would not be allowed to be cast if the player was in combat.

Resolved in co-operation with LordUsagi.