Fixed Aura overheals:
Every HoT (Heal over Time) aura appeared active and healed over time even though the targets
health was at maximum. Targets health did not increase over it's maximum value, but the visual
display of numbers appearing and disappearing showing the amount of health healed over the
characters avatar was active. The healing spell log should not be sent if there is no healing.

This was a visual bug caused by poorly designed code structure that handled spell healing.
To resolve this I've replaced this poor code structure with a bit less poor structure.

The result of all of this is a bit more flexibility with spell heal logs, less lines of code,
improved readability and a cleaner code setup.