Resolved issue #475 - Counterspell doesn't work:
Implemented a cleaner way to determine if a spell should be interrupted based on channel interrupt flags.
Evocation will now be interrupted by Counterspell as intended. All other spells that were affected by this
issue will now hopefully be rid of this problem.

This patch is property of LordUsagi and all credit goes to him.