Resolve issue #19 - Arcane Potency:
Arcane Potency is a talent that when 'proced', makes the caster enter
a 'clearcasting' state. Due to the spells intricacies it needed to be
handled in HandleDummyAuraProc. Previously it was not, and caused the
useful mage talent to become redundant.

Thanks to Cemak for posting the fix, LordUsagi for cleanups.

--HG--
branch : raw