[Core/Spells] Multiple spell changes including:
Use SpellCastResult to properly determine if a cast
is okay or not. This helps readability and hopefully
will improve the amount of code future developers can
understand.

Define and implement SPELL_ATTR_EX4_TRIGGERED which resolves
an issue with Water Elementals being able to cast Freeze from
across the map.