Core/Build Some build specific stuff
For some reason genrev.cmake needs to be invoked also
during the build process, but it will not see any variables.
Because of that I had to move find_package inside genrev.cmake.

Also I wrapped message()s with a check so no annoying messages will
be displayed during compiling (will only on configuring).

Hopefully we'll get around that in the future.

* genrev related issues should now be fixed:
* a little root cmake cleanup
* tell travis to build also extractors